add support for registering multiple ecovate::Listener instances
ecovate::Listener is a hack to allow us to listen for certain events
that the WebRTC stack doesn't ordinarly report to the application,
such as when the network layer is ready to send and when it's ready
for more data.  Ideally, we'd pass this listener all the way down the
stack, but that would require changing a lot of code, so we use a
hack: acquire a global lock and update a global reference to the
desired listener before calling PeerConnection::SetLocalDescription,
then clear the global reference and global lock.  That way, each
instance of PeerConnection gets a listener appropriate to that
listener, whereas the previous implementation only worked if there was
only one PeerConnection per process.

This commit also includes a couple of minor changes to ensure every
frame the application gives to the WebRTC stack makes it all the way
to the network (e.g. no frame dropping and no multithreading-induced
nondeterminism).  This makes writing a reliable integration test much
easier and aids debugging.